# backend/docs/openapi-governance.yaml
openapi: 3.1.0
info:
  title: DictionaryRoot Governed Platform API
  version: 0.2.0
  description: Authentication, account, governance, workflow, moderation, and audit additions to SourceRoot.
servers:
  - url: http://localhost:3000/api/v1
paths:
...
  /auth/email/start:
    post:
      summary: Request a one-time email sign-in or linking link
      responses: { '202': { description: Request accepted }, '429': { description: Rate limited } }
...
  /auth/identities/{identityId}:
    delete:
      summary: Unlink a non-active sign-in identity; the identity that created the current session cannot be removed
      parameters:
        - in: path
          name: identityId
          required: true
          schema: { type: string, format: uuid }
      responses: { '204': { description: Identity unlinked }, '409': { description: Active or final identity cannot be removed } }
...
components:
  securitySchemes:
    dictionaryRootSession:
      type: apiKey
      in: cookie
      name: dr_session
    csrfToken:
      type: apiKey
      in: header
      name: X-CSRF-Token